fallback to latest if getNPMVersion when network is down

+++ test/jest-global-setup.js
@@ -5,7 +5,8 @@ const fs = require("fs");
 
 module.exports = async function globalSetup() {
   const www = path.join(__dirname, "./www");
-  fs.unlinkSync(path.join(www, "package.json"));
+  if (fs.existsSync(path.join(www, "package.json")))
+    fs.unlinkSync(path.join(www, "package.json"));
   await startServer(www);
 
   await waitPort({
