+++ test/startServer.js
@@ -5,7 +5,8 @@ const fs = require("fs");
 
 async function globalSetup() {
   const www = path.join(__dirname, "./www");
-  fs.unlinkSync(path.join(www, "package.json"));
+  if (fs.existsSync(path.join(www, "package.json")))
+    fs.unlinkSync(path.join(www, "package.json"));
   await startServer(path.join(__dirname, "./www"));
 
   await waitPort({
